Add files via upload
I have used excel for transforming the data in the survey in a more understanding way. Further Analysis is done using pandas library of python and found total respondants , number of people answered the same answer. Lastly made a table with all the required information and exported the file.
Got the files from Shashank Kalanithi